+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - View: allowed to set DOM event handlers on a root element
 - CollectionView: added default implementation for "getHTML()" method returning simple "div" element
 - CollectionView: allowed not to specify EmptyView when it's not used
+- CollectionView: allowed to render Views inside of specified child Element
 - Documentation: improve example usage of "OsfObservable.retrigger()" method
 - Documentation: added "async" prefix to examples of View method usages
 - Documentation: fixed formatting in various code examples